ft_transcendence/backend/requirements.txt
ajsonrpc==1.2.0
anyio==4.1.0
asgiref==3.6.0
attrs==21.2.0
blinker==1.4
bottle==0.12.25
certifi==2020.6.20
channels==4.0.0
chardet==4.0.0
charset-normalizer==3.3.2
cli-helpers==2.2.1
click==8.1.7
colorama==0.4.6
configobj==5.0.6
cryptography==3.4.8
daphne==4.1.0
dbus-next==0.2.3
defer==1.0.4
distro==1.7.0
Django==4.2.10
djangorestframework==3.14.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
exceptiongroup==1.2.0
h11==0.14.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
introcs==1.3.1
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
marshmallow==3.20.2
more-itertools==8.10.0
mycli==1.24.3
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==23.2
pexpect==4.8.0
Pillow==9.0.1
prompt-toolkit==3.0.28
protobuf==3.12.4
ptyprocess==0.7.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
pyaes==1.6.1
pycairo==1.20.1
pyelftools==0.30
Pygments==2.11.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
python-dateutil==2.8.1
python-dotenv==0.19.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
SecretStorage==3.3.1
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.2
starlette==0.37.1
terminaltables==3.1.0
texttable==1.6.4
typing_extensions==4.8.0
urllib3==1.26.5
uvicorn==0.27.1
wadllib==1.3.6
wcwidth==0.2.5
wsproto==1.2.0
xdg==5
xlrd==1.2.0
zipp==1.0.0